eth/jwt: Add eth module check.
