+++ dex/networks/eth/params_test.go
@@ -6,8 +6,11 @@
 package eth
 
 import (
+	"io/fs"
 	"math"
 	"math/big"
+	"os"
+	"path/filepath"
 	"reflect"
 	"testing"
 )
@@ -94,6 +97,68 @@ func TestVersionedGases(t *testing.T) {
 		gas := RefundGas(tt.ver)
 		if gas != tt.expRefundGas {
 			t.Fatalf("wrong gas for refund, version %d. wanted %d, got %d", tt.ver, tt.expRefundGas, gas)
+		}
+	}
+}
+
+func TestFindJWTHex(t *testing.T) {
+	tDir := t.TempDir()
+	tFilePath := filepath.Join(tDir, "jwt.hex")
+	tests := []struct {
+		name, jwtFileContents, jwt, wantHex string
+		noPerms                             bool
+		wantErr                             bool
+	}{{
+		name:    "ok hex",
+		jwt:     "baadbeef",
+		wantHex: "baadbeef",
+	}, {
+		name:            "ok file",
+		jwt:             tFilePath,
+		jwtFileContents: "baadbeef\n",
+		wantHex:         "baadbeef",
+	}, {
+		name:    "not hex and no file",
+		jwt:     tFilePath,
+		wantErr: true,
+	}, {
+		name:            "not hex but cant read file",
+		jwt:             tFilePath,
+		jwtFileContents: "baadbeef\n",
+		noPerms:         true,
+		wantErr:         true,
+	}, {
+		name:            "not hex and bad hex in file",
+		jwt:             tFilePath,
+		jwtFileContents: "saadbeef\n",
+		wantErr:         true,
+	}}
+	for _, test := range tests {
+		if test.jwtFileContents != "" {
+			perms := fs.FileMode(0666)
+			if test.noPerms {
+				perms = 0000
+			}
+			err := os.WriteFile(tFilePath, []byte(test.jwtFileContents), perms)
+			if err != nil {
+				t.Fatal(err)
+			}
+		}
+		hex, err := FindJWTHex(test.jwt)
+		if test.jwtFileContents != "" {
+			os.Remove(tFilePath)
+		}
+		if test.wantErr {
+			if err == nil {
+				t.Fatalf("expected error for test %q", test.name)
+			}
+			continue
+		}
+		if err != nil {
+			t.Fatalf("unexpected error for test %q: %v", test.name, err)
+		}
+		if hex != test.wantHex {
+			t.Fatalf("wanted jwt %q but got %q for test %q", test.wantHex, hex, test.name)
 		}
 	}
 }
